# docs/openapi/innovation-sandbox-api.yaml
openapi: 3.0.0
info:
  title: Innovation Sandbox on AWS
  description: API documentation for the Innovation Sandbox on AWS solution.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "{apiBaseUrl}"
    description: Base URL used to make API requests
    variables:
      apiBaseUrl:
        default: https://example.com/api
        description: Base URL used to make API requests
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResponseBody:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - success
            - fail
            - error
      discriminator:
        propertyName: status
        mapping:
          success: "#/components/schemas/SuccessResponseBody"
          fail: "#/components/schemas/FailResponseBody"
          error: "#/components/schemas/ErrorResponseBody"
    SuccessResponseBody:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
        - type: object
          required:
            - status
            - data
          properties:
            status:
              type: string
              enum:
                - success
            data:
              type: object
    FailResponseBody:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
        - type: object
          required:
            - data
          properties:
            status:
              type: string
              enum:
                - fail
            data:
              type: object
    ErrorResponseBody:
      allOf:
        - $ref: "#/components/schemas/ResponseBody"
        - type: object
          required:
            - status
            - message
          properties:
            status:
              type: string
              enum:
                - error
            message:
              type: string
            code:
              type: number
            data:
              type: object
    PaginatedResults:
      type: object
      required:
        - items
      properties:
        items:
          type: array
        nextPageIdentifier:
          type: string
          example: "eyJsYXN0RXZhbHVhdGVkS2V5Ijp7InV1aWQiOnsicyI6IjEyMzQ1Njc4LTkwYWItY2RlZi0xMjM0LTU2Nzg5MGFiY2RlZiJ9fX0="
...
    BudgetThresholds:
      type: array
      items:
        type: object
        properties:
          dollarsSpent:
            type: number
            description: Threshold for budget notification
            example: 80
          action:
            type: string
            description: Action to take when threshold is reached
            enum: [ALERT, FREEZE_ACCOUNT]
            example: ALERT
    DurationThresholds:
      type: array
      items:
        type: object
        properties:
          hoursRemaining:
            type: number
            description: Threshold for durations notification
            example: 12
          action:
            type: string
            description: Action to take when threshold is reached
            enum: [ALERT, FREEZE_ACCOUNT]
            example: ALERT
...
    UserEmail:
      type: string
      format: email
      description: Email of the user who owns the lease
      example: "user@example.com"
...
    LeaseTemplateUuid:
      type: string
      format: uuid
      description: Unique identifier for the lease template used to create the lease
      example: "12345678-90ab-cdef-1234-567890abcdef"
    LeaseTemplateName:
      type: string
      description: Name of the lease template
      example: Example Template
    LeaseTemplateDescription:
      type: string
      description: Description of the lease template
      example: Example Template Description
    LeaseDurationInHours:
      type: integer
      description: Duration of the lease in hours
      example: 24
    MaxSpend:
      type: number
      description: Maximum spend allowed for the lease
      example: 100
...
    RequiresApproval:
      type: boolean
      description: Whether leases created from this template require approval
      example: true
    CreatedBy:
      type: string
      format: email
      description: Email of the user who created the template
      example: "admin@example.com"
    MetaData:
      type: object
      properties:
        createdTime:
          type: string
          format: date-time
          description: Time when the resource was created
          example: "2023-05-01T12:00:00Z"
        lastEditTime:
          type: string
          format: date-time
          description: Time when the resource was last edited
          example: "2023-05-01T12:00:00Z"
        schemaVersion:
          type: string
          description: Version of the schema used to create the resource
          example: 1
...
    LeaseTemplate:
      type: object
      required:
        - uuid
        - name
        - requiresApproval
        - createdBy
      properties:
        uuid:
          $ref: "#/components/schemas/LeaseTemplateUuid"
        name:
          $ref: "#/components/schemas/LeaseTemplateName"
        description:
          $ref: "#/components/schemas/LeaseTemplateDescription"
        requiresApproval:
          $ref: "#/components/schemas/RequiresApproval"
        createdBy:
          $ref: "#/components/schemas/UserEmail"
        maxSpend:
          $ref: "#/components/schemas/MaxSpend"
        budgetThresholds:
          $ref: "#/components/schemas/BudgetThresholds"
        leaseDurationInHours:
          $ref: "#/components/schemas/LeaseDurationInHours"
        durationThresholds:
          $ref: "#/components/schemas/DurationThresholds"
        meta:
          $ref: "#/components/schemas/MetaData"
...
  parameters:
    PageIdentifier:
      name: pageIdentifier
      in: query
      required: false
      schema:
        type: string
      description: Base64 encoded token for pagination
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
      description: Number of items to return per page
...
    LeaseTemplateId:
      name: leaseTemplateId
      description: UUID for lease template
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/LeaseTemplateUuid"
...
  responses:
    BadRequestResponse:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            status: fail
            data:
              errors: [{ "message": "example error message" }]
    UnauthorizedResponse:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            message: Unauthorized
    ResourceNotFoundResponse:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FailResponseBody"
          example:
            status: fail
            data:
              errors: [{ "message": "example error message" }]
...
    ErrorResponse:
      description: Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseBody"
          example:
            status: error
            message: "An unexpected error occurred"
paths:
...
  /leaseTemplates:
    get:
      tags:
        - leaseTemplates
      summary: Get paginated list of lease templates
      parameters:
        - $ref: "#/components/parameters/PageIdentifier"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PaginatedResults"
                          - type: object
                            properties:
                              result:
                                type: array
                                items:
                                  $ref: "#/components/schemas/LeaseTemplate"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags:
        - leaseTemplates
      summary: Create new lease template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
                - budgetConfig
                - durationConfig
                - requiresApproval
              properties:
                name:
                  $ref: "#/components/schemas/LeaseTemplateName"
                description:
                  $ref: "#/components/schemas/LeaseTemplateDescription"
                maxSpend:
                  $ref: "#/components/schemas/MaxSpend"
                leaseDurationInHours:
                  $ref: "#/components/schemas/LeaseDurationInHours"
                budgetThresholds:
                  $ref: "#/components/schemas/BudgetThresholds"
                durationThresholds:
                  $ref: "#/components/schemas/DurationThresholds"
                requiresApproval:
                  type: boolean
                  example: true
                  description: Whether leases created from this template require approval
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/LeaseTemplate"

        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /leaseTemplates/{leaseTemplateId}:
    parameters:
      - $ref: "#/components/parameters/LeaseTemplateId"
    get:
      tags:
        - leaseTemplates
      summary: Get lease template by id
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/LeaseTemplate"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    put:
      tags:
        - leaseTemplates
      summary: Update lease template by id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
                - requiresApproval
                - maxSpend
                - leaseDurationInHours
                - budgetThresholds
                - durationThresholds
                - createdBy
              properties:
                name:
                  type: string
                description:
                  type: string
                requiresApproval:
                  type: boolean
                maxSpend:
                  $ref: "#/components/schemas/MaxSpend"
                leaseDurationInHours:
                  $ref: "#/components/schemas/LeaseDurationInHours"
                budgetThresholds:
                  $ref: "#/components/schemas/BudgetThresholds"
                durationThresholds:
                  $ref: "#/components/schemas/DurationThresholds"
                createdBy:
                  $ref: "#/components/schemas/CreatedBy"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/LeaseTemplate"

        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags:
        - leaseTemplates
      summary: Delete lease template by id
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponseBody"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "403":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/ResourceNotFoundResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"

  # Account endpoints
...